WIP: Texture aligned with bodies
This took a lot of fussing with and I'm still not sure why things
are working the way they are. The x and y origin of the texture
entity is still not aligned to the box2d bodies. Should be centered,
however, at the moment it's located in the lower corner which makes
sense as that's where libGDX defaults it's origin to.